+++ yesod-core/test/RouteLeafCompile/MissingEndpoint.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE TypeFamilies #-}
+module MissingEndpoint where
+import Yesod.Core.RouteLeaf
+import YesodCoreTest.RouteLeaf.Foundation
+
+bad :: RouteLeaves AccountR -> Bool
+bad (LeafItemR _) = True