+++ yesod-core/test/RouteLeafCompile/MissingPolicy.hs
@@ -1,0 +1,11 @@
+{-# LANGUAGE TypeApplications #-}
+module MissingPolicy where
+import Yesod.Core (HandlerFor)
+import Yesod.Core.RouteLeaf
+import YesodCoreTest.RouteLeaf.Foundation
+import YesodCoreTest.RouteLeaf.Policy
+
+-- No concrete authorizers are imported. The generic dictionary must demand
+-- all endpoint-owner instances when it is instantiated here.
+bad :: HandlerFor LeafApp (Maybe ())
+bad = withRouteLeaves @AuthorizeRoute (const $ pure ())